feat: examples of analysis tools
