+++ Makefile
@@ -28,7 +28,7 @@ TERMINAL_CMAKE_CACHE := $(TERMINAL_BUILD_DIR)/CMakeCache.txt
 UNIT_TEST_CMAKE_CACHE := $(UNIT_TEST_BUILD_DIR)/CMakeCache.txt
 
 EXAMPLES_BUILD_DIR := build/examples
-EXAMPLES_BINARY := $(EXAMPLES_BUILD_DIR)/examples/genesys_fit_analysis
+EXAMPLES_BINARY := $(EXAMPLES_BUILD_DIR)/examples/genesys_examples_analysis_tools
 EXAMPLES_DATA := examples/data/sample_data.csv
 EXAMPLES_CMAKE_CACHE := $(EXAMPLES_BUILD_DIR)/CMakeCache.txt
 
@@ -131,7 +131,7 @@ $(EXAMPLES_CMAKE_CACHE):
 examples: $(EXAMPLES_CMAKE_CACHE)
 	cmake --build --preset examples -j$(JOBS)
 
-# Compila e executa o exemplo de análise de distribuições.
+# Compila e executa o exemplo de ferramentas de análise.
 # Passa o caminho do arquivo de dados a partir da raiz do projeto.
 run-examples: examples
 	$(EXAMPLES_BINARY) $(EXAMPLES_DATA)
